Gcc versiyonu kaynaklı ortaya çıkan hata giderildi.

+++ src/C++/FieldConvertors.h
@@ -37,7 +37,7 @@ namespace FIX
 typedef int signed_int;
 typedef unsigned int unsigned_int;
 
-#define UNSIGNED_VALUE_OF( x ) unsigned_int( x < 0 ? -x : x )
+#define UNSIGNED_VALUE_OF( x ) unsigned_int( x < 0 ? -(long long)x : x )
 
 #define IS_SPACE( x ) ( x == ' ' )
 #define IS_DIGIT( x ) ( unsigned_int( x - '0' ) < 10 )
@@ -187,7 +187,7 @@ struct IntConvertor
     } while ( ++str != end );
 
     if( isNegative )
-      x = -x;
+      x = -(long long)x;
 
     result = x;
     return true;
